:memo: :zap: applied requested changes
